+++ internal/pkg/agent/vault/seed_nofips.go
@@ -6,16 +6,30 @@
 
 package vault
 
-import "errors"
+import (
+	"fmt"
+	"io/fs"
+	"os"
+	"path/filepath"
 
-func getSeed(path string) ([]byte, int, error) {
-	b, errV1 := getSeedV1(path)
-	if errV1 == nil {
-		return b, saltSizeV1, nil
+	"github.com/elastic/elastic-agent/internal/pkg/agent/vault/aesgcm"
+)
+
+func getSeedV1(path string) ([]byte, error) {
+	fp := filepath.Join(path, seedFile)
+
+	b, err := os.ReadFile(fp)
+	if err != nil {
+		return nil, fmt.Errorf("could not read seed file: %w", err)
 	}
-	b, saltSize, errV2 := getSeedV2(path)
-	if errV2 == nil {
-		return b, saltSize, nil
+
+	// return fs.ErrNotExist if invalid length of bytes returned
+	if len(b) != int(aesgcm.AES256) {
+		return nil, fmt.Errorf("invalid seed length, expected: %v, got: %v: %w", int(aesgcm.AES256), len(b), fs.ErrNotExist)
 	}
-	return nil, 0, errors.Join(errV1, errV2)
+	return b, nil
+}
+
+func checkSalt(_ int) error {
+	return nil
 }
